apicrud-golang improved, file bin deleted, .env.example file mas entendible

+++ user.go
@@ -2,9 +2,12 @@ package main
 
 import (
 	"encoding/json"
+	"fmt"
 	"net/http"
+	"strconv"
 
 	"github.com/gorilla/mux"
+	"gorm.io/gorm"
 )
 
 func GetUsers(w http.ResponseWriter, r *http.Request) {
@@ -14,11 +17,28 @@ func GetUsers(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(users)
 }
 
-func GetUser(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
+func GetUserByID(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
+	idClean, err := strconv.Atoi(params["id"])
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	if idClean == 0 {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
 	var user User
-	DB.First(&user, params["id"])
+	if err := DB.First(&user, params["id"]).Error; err != nil {
+		if err == gorm.ErrRecordNotFound {
+			w.WriteHeader(http.StatusNotFound)
+			return
+		}
+		fmt.Println(err.Error())
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
 	json.NewEncoder(w).Encode(user)
 }
 
@@ -26,24 +46,50 @@ func CreateUser(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	var user User
 	json.NewDecoder(r.Body).Decode(&user)
+	if user.Name == "" || user.Address == "" || user.Age == 0 || user.Email == "" || user.Phone == "" || user.Username == "" {
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
 	DB.Create(&user)
 	json.NewEncoder(w).Encode(user)
 }
 
-func UpdateUser(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
+func UpdateUserByID(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
+	idClean, err := strconv.Atoi(params["id"])
+	if err != nil {
+		fmt.Println(err.Error())
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	if idClean == 0 {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
 	var user User
 	DB.First(&user, params["id"])
 	json.NewDecoder(r.Body).Decode(&user)
+	if user.Name == "" || user.Address == "" || user.Age == 0 || user.Email == "" || user.Phone == "" || user.Username == "" {
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
 	DB.Save(&user)
 	json.NewEncoder(w).Encode(user)
 }
 
-func DeleteUser(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
+func DeleteUserByID(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
+	idClean, err := strconv.Atoi(params["id"])
+	if err != nil {
+		fmt.Println(err.Error())
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	if idClean == 0 {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
 	var user User
 	DB.Delete(&user, params["id"])
-	json.NewEncoder(w).Encode(user)
 }
